% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_assets.R
\name{app_assets}
\alias{app_assets}
\alias{app_styles}
\alias{compile_styles}
\alias{app_js}
\alias{app_favicon}
\alias{add_resource_path}
\alias{add_external_resources}
\title{App Assets}
\usage{
app_styles(path = "www/styles.min.css")

compile_styles(
  input_scss = "www/styles/scss/index.scss",
  output_css = "www/styles/css/styles.min.css",
  minify = TRUE,
  ...
)

app_js(path = "www/scripts/js")

app_favicon(path = "www/favicon.ico")

add_resource_path(prefix = "www", path = pkg_sys("www"), overwrite = TRUE)

add_external_resources()
}
\arguments{
\item{path}{Path to the asset file, resource directory, favicon, etc.
Defaults are provided for each function.}

\item{minify}{Logical. Should the CSS be minified? Defaults to \code{TRUE}.}

\item{prefix}{Prefix for the resource path. Defaults to \code{"www"}.}

\item{overwrite}{Logical. Should the resource path be overwritten if it already exists?
Defaults to \code{TRUE}.}

\item{scss_path}{Path to the SCSS file. Defaults to \code{"www/styles/scss/index.scss"}.}

\item{css_path}{Path to the CSS file. Defaults to \code{"www/styles/css/styles.min.css"}.}
}
\value{
\itemize{
\item \code{app_styles()}: HTML \verb{<head>} tag with a link to the CSS file.
\item \code{compile_styles()}: Compiled CSS file.
\item \code{app_favicon()}: HTML \verb{<head>} tag with a link to the favicon.
\item \code{add_resource_path()}: \code{NULL}, used for its side effect of adding a resource path.
\item \code{add_external_resources()}: HTML \verb{<head>} tags with links to external resources.
}
}
\description{
Various functions for including, bundling, and managing Shiny App static assets:
\itemize{
\item \code{app_styles()}: Add CSS styles to the Shiny application.
\item \code{compile_styles()}: Compile SCSS styles into CSS for the Shiny application.
\item \code{app_favicon()}: Add a favicon to the Shiny application.
\item \code{add_resource_path()}: Add a resource path to the Shiny application.
\item \code{add_external_resources()}: Add external resources to the Shiny application.
}
}
\seealso{
Other App: 
\code{\link{app_server}()},
\code{\link{app_ui}()},
\code{\link{run_app}()}
}
\concept{App}
