% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{App Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Default shiny server arguments.}
}
\value{
Shiny Server
}
\description{
App Server
}
\seealso{
Other App: 
\code{\link{app_assets}},
\code{\link{app_ui}()},
\code{\link{run_app}()}
}
\concept{App}
