% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_theme.R
\name{app_theme}
\alias{app_theme}
\title{App Theme}
\usage{
app_theme()
}
\value{
A Bootstrap theme object.
}
\description{
Custom Bootstrap theme for the Shiny app.
}
\details{
This function defines a custom Bootstrap theme for the Shiny app
using \code{\link[bslib:bs_theme]{bslib::bs_theme()}}.
}
