% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Core Shiny App User Interface}
\usage{
app_ui(req)
}
\arguments{
\item{req}{(Internal) The initial HTTP request object.}
}
\value{
\code{\link[htmltools:tagList]{htmltools::tagList()}} containing the core user interface elements.
}
\description{
This function generates the core user interface for the GMH DataHub Shiny app.
}
\seealso{
\code{\link[=app_server]{app_server()}}

Other App: 
\code{\link{app_assets}},
\code{\link{app_server}()},
\code{\link{run_app}()}
}
\concept{App}
