% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cli_msg}
\alias{cli_msg}
\title{Messages}
\usage{
cli_msg(msg, type = "info", .log = FALSE)
}
\arguments{
\item{msg}{Message to display.}

\item{type}{Type of message: "info", "warning", or "error".}
}
\description{
This function is a wrapper for \code{\link[cli:cli_alert]{cli::cli_alert()}} to provide consistent
console messaging.
}
\seealso{
\code{\link[cli:cli_alert]{cli::cli_alert()}}
}
