% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_connect}
\alias{db_connect}
\title{Connect to GMH Database}
\usage{
db_connect(db_config = get_db_config(), user_id = NULL)
}
\arguments{
\item{db_config}{A named list containing the required database connection
parameters. The list should contain the following elements: \code{host},
\code{port}, \code{dbname}, \code{user}, and \code{password}. This list is typically
retrieved via a \code{config.yml} configuration file or via the
\code{\link[=get_db_config]{get_db_config()}} function (the default).}

\item{user_id}{An integer representing the user id to use for the database
connection. If not provided, the function will attempt to retrieve the
user id from the shiny session object. If the function is not running
inside a shiny app, the default user id of \code{0} will be used.}
}
\value{
A \code{pool} object representing the connection pool to the GMH
PostgreSQL database.
}
\description{
This function establishes a connection pool to the GMH PostgreSQL database
hosted in Google Cloud SQL using the \code{pool} package. Additionally,
the function will setup automatic cleanup of the connection pool when
the R or shiny session ends.
}
