% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{elapsed_months}
\alias{elapsed_months}
\title{Elapsed Months}
\usage{
elapsed_months(start_date, end_date)
}
\arguments{
\item{start_date}{Date representing the start of a period.}

\item{end_date}{Date representing the end of a period.}
}
\value{
The number of months elapsed between the two dates.
}
\description{
Derive the number of months elapsed between two dates.
}
