% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_errors.R
\name{entrata_errors}
\alias{entrata_errors}
\alias{entrata_req_error}
\alias{entrata_resp_is_error}
\alias{entrata_resp_error_body}
\title{Entrata Error Handling}
\usage{
entrata_req_error(
  req,
  is_error = entrata_resp_is_error,
  body = entrata_resp_error_body
)

entrata_resp_is_error(resp)

entrata_resp_error_body(resp)
}
\arguments{
\item{req}{\code{httr2::request} object.}

\item{is_error}{A predicate function that takes a single argument (the
response) and returns \code{TRUE} or \code{FALSE} indicating whether or not an
R error should be signalled.}

\item{body}{A callback function that takes a single argument (the response)
and returns a character vector of additional information to include in the
body of the error. This vector is passed along to the \code{message} argument
of \code{\link[rlang:abort]{rlang::abort()}} so you can use any formatting that it supports.}

\item{resp}{\code{httr2::response} object.}
}
\value{
\itemize{
\item \code{entrata_req_error()} returns the altered request object.
\item \code{entrata_resp_is_error()} returns a \code{logical} indicating if the response is an error.
\item \code{entrata_resp_error_body()} returns a \code{list} with the error code and message.
}
}
\description{
Functions for handling errors from the Entrata API.
}
\details{
\itemize{
\item \code{entrata_req_error()} alters the Entrata request object with a custom error handler.
\item \code{entrata_resp_is_error()} checks if the response from the Entrata API is an error.
\item \code{entrata_resp_error_body()} extracts the error body and message from the response.
}
}
\seealso{
\code{\link[httr2:req_error]{httr2::req_error()}}, \code{\link[=entrata_request]{entrata_request()}}
}
