% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_leases.R
\name{entrata_leases}
\alias{entrata_leases}
\alias{construct_entrata_leases_request_method_params}
\alias{entrata_resp_parse_leases}
\alias{entrata_resp_parse_lease_intervals}
\alias{entrata_resp_parse_lease_customers}
\alias{entrata_resp_parse_lease_scheduled_charges}
\alias{entrata_resp_parse_lease_activities}
\title{Entrata Leases}
\usage{
entrata_leases(property_id, request_id = NULL, ...)

construct_entrata_leases_request_method_params(property_id, ...)

entrata_resp_parse_leases(resp)

entrata_resp_parse_lease_intervals(resp_parsed)

entrata_resp_parse_lease_customers(resp_parsed)

entrata_resp_parse_lease_scheduled_charges(resp_parsed)

entrata_resp_parse_lease_activities(resp_parsed)
}
\arguments{
\item{property_id}{The ID of the property.}

\item{request_id}{The ID of the request.}

\item{...}{Additional parameters passed to the API endpoint.}
}
\value{
API response
}
\description{
Builds an API request for Entrata's \verb{/leases} endpoint, \code{getLeases} method.
}
\details{
The \code{getLeases} method retrieves a list of leases for a property.
}
\section{Additional Parameters}{


The following additional parameters can be passed to the API request:
\itemize{
\item \code{applicationId} (integer) - The ID of the application.
\item \code{customerId} (integer) - The ID of the customer.
\item \code{leaseStatusTypeIds} (integer) - The ID of the lease status type.
\item \code{leaseIds} (integer) - The ID of the lease.
\item \code{scheduledArCodeIds} (integer) - The ID of the scheduled AR code.
\item \code{unitNumber} (string) - The unit number.
\item \code{buildingName} (string) - The building name.
\item \code{moveInDateFrom} (string) - The move-in date from.
\item \code{moveInDateTo} (string) - The move-in date to.
\item \code{leaseExpiringDateFrom} (string) - The lease expiring date from.
\item \code{leaseExpiringDateTo} (string) - The lease expiring date to.
\item \code{moveOutDateFrom} (string) - The move-out date from.
\item \code{moveOutDateTo} (string) - The move-out date to.
\item \code{includeOtherIncomeLeases} (boolean) - Include other income leases.
\item \code{residentFriendlyMode} (boolean) - Resident friendly mode.
\item \code{includeLeaseHistory} (boolean) - Include lease history.
\item \code{includeArTransactions} (boolean) - Include AR transactions.
}

Additionally, the following pagination parameters can be passed as \code{...} to
the function and will be included in the API request:
\itemize{
\item \code{pageNumber} (integer) - The page number.
\item \code{perPage} (integer) - The number of items per page.
\item \code{includePaginationLinks} (boolean) - Include pagination links.
}
}

