% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logger.R
\name{entrata_logger_init}
\alias{entrata_logger_init}
\title{Intialize Entrata Logger}
\usage{
entrata_logger_init(log_file = "entrata.log", ...)
}
\arguments{
\item{log_file}{The name of the log file to write to. Default is \code{"entrata.log"}.}

\item{...}{Additional arguments to pass to the logger. Currently not in use.}
}
\description{
Initialize the logger for the Entrata API.

By default, the logger will write to a file named \code{"entrata.log"} in the
current working directory and leverage the following logging setup:
\itemize{
\item Formatter: \code{logger::formatter_glue}
\item Appender: \code{logger::appender_tee}
\item Layout: \code{logger::layout_glue_colors}
\item Threshold: \code{logger::INFO}
}
}
