% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_queue.R
\name{entrata_queue}
\alias{entrata_queue}
\title{Entrata \verb{/queue} Endpoint}
\usage{
entrata_queue(queue_id, entrata_config = get_entrata_config())
}
\arguments{
\item{queue_id}{The unique identifier for the queue item.}

\item{entrata_config}{A list containing the necessary configuration}
}
\value{
A list containing the response body and status code.
}
\description{
The \verb{/queue} endpoint provides a mechanism for retrieving previously
created queue items. Queue items are created when a user requests a
report that takes a long time to generate. The user is then able to
check the status of the report by querying the queue endpoint with the
\code{queueId}.
}
