% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_request.R
\name{entrata_req_auth}
\alias{entrata_req_auth}
\title{Entrata Request Authentication}
\usage{
entrata_req_auth(req, username = NULL, password = NULL)
}
\arguments{
\item{req}{An \code{\link[httr2:request]{httr2::request()}} object}

\item{username}{A character string representing the username for the Entrata API}

\item{password}{A character string representing the password for the Entrata API}
}
\value{
The modified request object with authentication headers.
}
\description{
Adds basic authentication to the Entrata request object using provided credentials.
}
