% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_request.R
\name{entrata_req_body}
\alias{entrata_req_body}
\title{Entrata Request Body}
\usage{
entrata_req_body(
  req,
  id = NULL,
  method = NULL,
  version = NULL,
  params = list()
)
}
\arguments{
\item{req}{An httr2 request object}

\item{id}{Request ID}

\item{method}{Request Endpoint Method (not HTTP method)}

\item{version}{Request Endpoint Method Version}

\item{params}{Request Endpoint Method Parameters}
}
\value{
The modified request object with the new request body options
}
\description{
Modifies the Entrata request object with the provided request body options.

Note that the endpoint should be set before attempting to use this function
to get the best results.
}
