% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_request.R
\name{entrata_req_headers}
\alias{entrata_req_headers}
\title{Entrata Request Headers}
\usage{
entrata_req_headers(req, headers)
}
\arguments{
\item{req}{An \code{\link[httr2:request]{httr2::request()}} object.}

\item{headers}{A named list of headers to set on the request object.}
}
\value{
The modified request object with the new headers set.
}
\description{
Modifies the Entrata request object with the provided headers.
}
