% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logger.R
\name{entrata_req_log}
\alias{entrata_req_log}
\title{Entrata Request Logger}
\usage{
entrata_req_log(req)
}
\arguments{
\item{req}{The \code{httr2::request()} object.}
}
\description{
Log the details of an HTTP request to the Entrata API.
}
