% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_request.R
\name{entrata_req_pagination}
\alias{entrata_req_pagination}
\title{Entrata Request Pagination}
\usage{
entrata_req_pagination(
  req,
  page_number = 1,
  per_page = 500,
  include_pagination_links = FALSE
)
}
\arguments{
\item{req}{An \code{\link[httr2:request]{httr2::request()}} object.}

\item{page_number}{Pagination page number. Added to the request's URL query
parameters. Default is \code{1} (i.e. \verb{/?page_no=1}).}

\item{per_page}{Number of items per page. Added to the request's URL query
parameters. Default is \code{500} (i.e. \verb{/?page_no=1&per_page=500}).}

\item{include_pagination_links}{Logical value to include pagination links in
the response. Default is \code{FALSE}. Alters the request's headers by injecting
the \code{X-Send-Pagination-Links} header.}
}
\value{
The modified request object with pagination headers and query parameters.
}
\description{
Many of the Entrata API endpoints support pagination. This function
adds pagination related headers and query parameters to the request object.
}
