% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_request.R
\name{entrata_req_user_agent}
\alias{entrata_req_user_agent}
\title{Entrata Request User Agent}
\usage{
entrata_req_user_agent(req, string)
}
\arguments{
\item{req}{An httr2 request object}

\item{string}{A character string representing the user agent to set on the request
object. If not provided, will use the default user agent set in the Entrata
configuration or the package version and URL.}
}
\value{
The modified request object with the new user agent string
}
\description{
Modifies the Entrata request object with the provided user agent string.
}
