% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_validation.R
\name{entrata_req_validate}
\alias{entrata_req_validate}
\alias{entrata_resp_validate}
\title{Entrata Request and Response Validation}
\usage{
entrata_req_validate(req, ...)

entrata_resp_validate(
  resp,
  arg = rlang::caller_arg(resp),
  call = rlang::caller_env()
)
}
\arguments{
\item{req}{A request object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
The request object.
}
\description{
Entrata API request and response validation functions.
}
