% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_request.R
\name{entrata_request}
\alias{entrata_request}
\title{Initialize an Entrata API Request}
\usage{
entrata_request(entrata_config = get_entrata_config())
}
\arguments{
\item{entrata_config}{A list containing Entrata configuration
(Base URL, Username and Password, etc.). By default will call
\code{\link[=get_entrata_config]{get_entrata_config()}} to retrieve the configuration.}
}
\value{
An \code{\link[httr2:request]{httr2::request()}} object with Entrata-specific configurations.
}
\description{
Sets up a base request object with core Entrata API configurations.
}
