% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_request.R
\name{entrata_request_modifiers}
\alias{entrata_request_modifiers}
\title{Entrata Request Modifiers}
\description{
These functions are used to modify the Entrata request object before
sending the request to the API. These functions are used to set the
endpoint path, Entrata internal endpoint method (name, version, and params),
additional headers, authentication, request body, error handling, logging,
and more.
\itemize{
\item \code{\link[=entrata_req_endpoint]{entrata_req_endpoint()}} - Set the endpoint path for the request
\item \code{\link[=entrata_req_body]{entrata_req_body()}} - Set the request body options
\item \code{\link[=entrata_req_auth]{entrata_req_auth()}} - Add basic authentication to the request
\item \code{\link[=entrata_req_user_agent]{entrata_req_user_agent()}} - Set the user agent for the request
\item \code{\link[=entrata_req_retry]{entrata_req_retry()}} - Set the retry policy for the request
\item \code{\link[=entrata_req_log]{entrata_req_log()}} - Set the logging policy for the request
\item \code{\link[=entrata_req_cache]{entrata_req_cache()}} - Set the caching policy for the request
\item \code{\link[=entrata_req_hash]{entrata_req_hash()}} - Generate a hash key for the request
}
}
\seealso{
\code{\link[=entrata_request]{entrata_request()}}
\code{\link[=entrata_req_endpoint]{entrata_req_endpoint()}}, \code{\link[=entrata_req_body]{entrata_req_body()}}, \code{\link[=entrata_req_auth]{entrata_req_auth()}},
\code{\link[=entrata_req_user_agent]{entrata_req_user_agent()}}, \code{\link[=entrata_req_headers]{entrata_req_headers()}}, \code{\link[=entrata_req_config]{entrata_req_config()}},
\code{\link[=entrata_req_log]{entrata_req_log()}}
}
