% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logger.R
\name{entrata_resp_log}
\alias{entrata_resp_log}
\title{Entrata Response Logger}
\usage{
entrata_resp_log(resp)
}
\arguments{
\item{resp}{The \code{httr2::response()} object.}
}
\description{
Log the details of an HTTP response from the Entrata API.
}
