% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{extract_date}
\alias{extract_date}
\title{Extract Date from String}
\usage{
extract_date(string)
}
\arguments{
\item{string}{The string to extract the date from.}
}
\value{
The extracted date.
}
\description{
Extract Date from String
}
