% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_formats.R
\name{extract_phone_number}
\alias{extract_phone_number}
\title{Extract Phone Number}
\usage{
extract_phone_number(input_string)
}
\arguments{
\item{input_string}{A string containing a phone number}
}
\value{
A string formatted for a "tel://" href
}
\description{
Extracts the phone number from a string and formats it for a "tel://" href.
}
