% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_formats.R
\name{format_phone_number}
\alias{format_phone_number}
\title{Formats a phone number}
\usage{
format_phone_number(phone)
}
\arguments{
\item{phone}{Phone number to format}
}
\value{
String of formatted phone number
}
\description{
This function formats phone numbers, and by default uses the syntax:
\verb{+1 (555) 555-5555}.
}
