% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_defaults.R
\name{get_default_method_version}
\alias{get_default_method_version}
\title{Get Default Entrata Method Version}
\usage{
get_default_method_version(endpoint, method)
}
\arguments{
\item{endpoint}{The Entrata endpoint}

\item{method}{The Entrata method}
}
\value{
The default version of the Entrata method
}
\description{
Get Default Entrata Method Version
}
