% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_github.R
\name{get_gh_repo_last_modified}
\alias{get_gh_repo_last_modified}
\title{GitHub Repository Last Modified Date}
\usage{
get_gh_repo_last_modified(repo, format = "\%Y-\%m-\%d")
}
\arguments{
\item{repo}{A character string of the GitHub repository in the format 'user/repo'.}

\item{format}{A character string specifying the format of the date to return.}
}
\value{
A character string of the last modified date of the GitHub repository.
}
\description{
Get the last modified date of a GitHub repository.
}
\examples{
get_gh_repo_last_modified("noclocks/gmhleasr")
}
