% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmhcommunities-package.R
\docType{package}
\name{gmhcommunities-package}
\alias{gmhcommunities}
\alias{gmhcommunities-package}
\title{gmhcommunities: GMH Communities Data Hub Platform}
\description{
Provides a suite of tools for working with the GMH Communities Data Hub Platform.
}
\details{
\code{gmhcommunities} is a comprehensive R package encapsulating various R
functions and workflows related to GMH Communities and their data
provider Entrata.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/noclocks/gmhdatahub}
  \item Report bugs at \url{https://github.com/noclocks/gmhdatahub/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

Authors:
\itemize{
  \item Patrick Howard \email{patrick.howard@noclocks.dev} [reviewer]
}

Other contributors:
\itemize{
  \item No Clocks, LLC \email{team@noclocks.dev} [copyright holder, funder]
}

}
\keyword{internal}
