% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dashboard.R
\name{mod_dashboard}
\alias{mod_dashboard}
\alias{mod_dashboard_ui}
\alias{mod_dashboard_server}
\alias{mod_dashboard_demo}
\title{Dashboard Shiny Module}
\usage{
mod_dashboard_ui(id)

mod_dashboard_server(id)

mod_dashboard_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_dashboard_ui()}: UI HTML Output.
\item \code{mod_dashboard_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny module that displays a dashboard with key metrics and visualizations.
\itemize{
\item \code{mod_dashboard_ui()}: User interface
\item \code{mod_dashboard_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_dashboard_demo()
}
}
