% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_survey_fees.R
\name{mod_market_survey_fees}
\alias{mod_market_survey_fees}
\alias{mod_market_survey_fees_ui}
\alias{mod_market_survey_fees_server}
\alias{mod_market_survey_fees_demo}
\title{market_survey_fees Shiny Module}
\usage{
mod_market_survey_fees_ui(id)

mod_market_survey_fees_server(id)

mod_market_survey_fees_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_market_survey_fees_ui()}: UI HTML Output.
\item \code{mod_market_survey_fees_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_market_survey_fees_ui()}: User interface
\item \code{mod_market_survey_fees_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_market_survey_fees_demo()
}
}
