% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_survey_hours.R
\name{mod_market_survey_hours}
\alias{mod_market_survey_hours}
\alias{mod_market_survey_hours_ui}
\alias{mod_market_survey_hours_server}
\alias{mod_market_survey_hours_demo}
\title{market_survey_hours Shiny Module}
\usage{
mod_market_survey_hours_ui(id)

mod_market_survey_hours_server(id)

mod_market_survey_hours_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_market_survey_hours_ui()}: UI HTML Output.
\item \code{mod_market_survey_hours_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_market_survey_hours_ui()}: User interface
\item \code{mod_market_survey_hours_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_market_survey_hours_demo()
}
}
