% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_survey_notes.R
\name{mod_market_survey_notes}
\alias{mod_market_survey_notes}
\alias{mod_market_survey_notes_ui}
\alias{mod_market_survey_notes_server}
\alias{mod_market_survey_notes_demo}
\title{market_survey_notes Shiny Module}
\usage{
mod_market_survey_notes_ui(id)

mod_market_survey_notes_server(id)

mod_market_survey_notes_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_market_survey_notes_ui()}: UI HTML Output.
\item \code{mod_market_survey_notes_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_market_survey_notes_ui()}: User interface
\item \code{mod_market_survey_notes_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_market_survey_notes_demo()
}
}
