% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_survey_parking.R
\name{mod_market_survey_parking}
\alias{mod_market_survey_parking}
\alias{mod_market_survey_parking_ui}
\alias{mod_market_survey_parking_server}
\alias{mod_market_survey_parking_demo}
\title{market_survey_parking Shiny Module}
\usage{
mod_market_survey_parking_ui(id)

mod_market_survey_parking_server(id)

mod_market_survey_parking_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_market_survey_parking_ui()}: UI HTML Output.
\item \code{mod_market_survey_parking_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_market_survey_parking_ui()}: User interface
\item \code{mod_market_survey_parking_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_market_survey_parking_demo()
}
}
