% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_survey_property_summary.R
\name{mod_market_survey_property_summary}
\alias{mod_market_survey_property_summary}
\alias{mod_market_survey_property_summary_ui}
\alias{mod_market_survey_property_summary_server}
\alias{mod_market_survey_property_summary_demo}
\title{market_survey_property_summary Shiny Module}
\usage{
mod_market_survey_property_summary_ui(id)

mod_market_survey_property_summary_server(
  id,
  selected_property = NULL,
  pool = NULL,
  property_data = NULL
)

mod_market_survey_property_summary_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_market_survey_property_summary_ui()}: UI HTML Output.
\item \code{mod_market_survey_property_summary_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_market_survey_property_summary_ui()}: User interface
\item \code{mod_market_survey_property_summary_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_market_survey_property_summary_demo()
}
}
