% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_survey_rents.R
\name{mod_market_survey_rents}
\alias{mod_market_survey_rents}
\alias{mod_market_survey_rents_ui}
\alias{mod_market_survey_rents_server}
\alias{mod_market_survey_rents_demo}
\title{market_survey_rents Shiny Module}
\usage{
mod_market_survey_rents_ui(id)

mod_market_survey_rents_server(
  id,
  pool,
  selected_property = NULL,
  selected_leasing_week = NULL
)

mod_market_survey_rents_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_market_survey_rents_ui()}: UI HTML Output.
\item \code{mod_market_survey_rents_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_market_survey_rents_ui()}: User interface
\item \code{mod_market_survey_rents_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_market_survey_rents_demo()
}
}
