% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_survey_short_term_leases.R
\name{mod_market_survey_short_term_leases}
\alias{mod_market_survey_short_term_leases}
\alias{mod_market_survey_short_term_leases_ui}
\alias{mod_market_survey_short_term_leases_server}
\alias{mod_market_survey_short_term_leases_demo}
\title{market_survey_short_term_leases Shiny Module}
\usage{
mod_market_survey_short_term_leases_ui(id)

mod_market_survey_short_term_leases_server(id)

mod_market_survey_short_term_leases_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_market_survey_short_term_leases_ui()}: UI HTML Output.
\item \code{mod_market_survey_short_term_leases_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_market_survey_short_term_leases_ui()}: User interface
\item \code{mod_market_survey_short_term_leases_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_market_survey_short_term_leases_demo()
}
}
