% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_survey_utilities.R
\name{mod_market_survey_utilities}
\alias{mod_market_survey_utilities}
\alias{mod_market_survey_utilities_ui}
\alias{mod_market_survey_utilities_server}
\alias{mod_market_survey_utilities_demo}
\title{market_survey_utilities Shiny Module}
\usage{
mod_market_survey_utilities_ui(id)

mod_market_survey_utilities_server(id)

mod_market_survey_utilities_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_market_survey_utilities_ui()}: UI HTML Output.
\item \code{mod_market_survey_utilities_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_market_survey_utilities_ui()}: User interface
\item \code{mod_market_survey_utilities_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_market_survey_utilities_demo()
}
}
