% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pre_lease_details.R
\name{mod_pre_lease_details}
\alias{mod_pre_lease_details}
\alias{mod_pre_lease_details_ui}
\alias{mod_pre_lease_details_server}
\alias{mod_pre_lease_details_demo}
\title{pre_lease_details Shiny Module}
\usage{
mod_pre_lease_details_ui(id)

mod_pre_lease_details_server(id)

mod_pre_lease_details_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_pre_lease_details_ui()}: UI HTML Output.
\item \code{mod_pre_lease_details_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_pre_lease_details_ui()}: User interface
\item \code{mod_pre_lease_details_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_pre_lease_details_demo()
}
}
