% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_properties.R
\name{mod_properties}
\alias{mod_properties}
\alias{mod_properties_ui}
\alias{mod_properties_server}
\alias{mod_properties_demo}
\title{properties Shiny Module}
\usage{
mod_properties_ui(id)

mod_properties_server(id)

mod_properties_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_properties_ui()}: UI HTML Output.
\item \code{mod_properties_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_properties_ui()}: User interface
\item \code{mod_properties_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_properties_demo()
}
}
