% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_property_units.R
\name{mod_property_units}
\alias{mod_property_units}
\alias{mod_property_units_ui}
\alias{mod_property_units_server}
\alias{mod_property_units_demo}
\title{property_units Shiny Module}
\usage{
mod_property_units_ui(id)

mod_property_units_server(id)

mod_property_units_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_property_units_ui()}: UI HTML Output.
\item \code{mod_property_units_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_property_units_ui()}: User interface
\item \code{mod_property_units_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_property_units_demo()
}
}
