% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_insights.R
\name{mod_survey_insights}
\alias{mod_survey_insights}
\alias{mod_survey_insights_ui}
\alias{mod_survey_insights_server}
\alias{mod_survey_insights_demo}
\title{survey_insights Shiny Module}
\usage{
mod_survey_insights_ui(id)

mod_survey_insights_server(id)

mod_survey_insights_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_survey_insights_ui()}: UI HTML Output.
\item \code{mod_survey_insights_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_survey_insights_ui()}: User interface
\item \code{mod_survey_insights_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_survey_insights_demo()
}
}
