% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_insights_results.R
\name{mod_survey_insights_results}
\alias{mod_survey_insights_results}
\alias{mod_survey_insights_results_ui}
\alias{mod_survey_insights_results_server}
\alias{mod_survey_insights_results_demo}
\title{survey_insights_results Shiny Module}
\usage{
mod_survey_insights_results_ui(id)

mod_survey_insights_results_server(id)

mod_survey_insights_results_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_survey_insights_results_ui()}: UI HTML Output.
\item \code{mod_survey_insights_results_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_survey_insights_results_ui()}: User interface
\item \code{mod_survey_insights_results_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_survey_insights_results_demo()
}
}
