% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_insights_swot.R
\name{mod_survey_insights_swot}
\alias{mod_survey_insights_swot}
\alias{mod_survey_insights_swot_ui}
\alias{mod_survey_insights_swot_server}
\alias{mod_survey_insights_swot_demo}
\title{survey_insights_swot Shiny Module}
\usage{
mod_survey_insights_swot_ui(id)

mod_survey_insights_swot_server(id)

mod_survey_insights_swot_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_survey_insights_swot_ui()}: UI HTML Output.
\item \code{mod_survey_insights_swot_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_survey_insights_swot_ui()}: User interface
\item \code{mod_survey_insights_swot_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_survey_insights_swot_demo()
}
}
