% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_insights_trends.R
\name{mod_survey_insights_trends}
\alias{mod_survey_insights_trends}
\alias{mod_survey_insights_trends_ui}
\alias{mod_survey_insights_trends_server}
\alias{mod_survey_insights_trends_demo}
\title{survey_insights_trends Shiny Module}
\usage{
mod_survey_insights_trends_ui(id)

mod_survey_insights_trends_server(id)

mod_survey_insights_trends_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_survey_insights_trends_ui()}: UI HTML Output.
\item \code{mod_survey_insights_trends_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_survey_insights_trends_ui()}: User interface
\item \code{mod_survey_insights_trends_server()}: Server logic
}
}
\examples{
if (interactive()) {
  mod_survey_insights_trends_demo()
}
}
