% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_title.R
\name{mod_title}
\alias{mod_title}
\alias{mod_title_ui}
\alias{mod_title_server}
\title{Title Module}
\usage{
mod_title_ui(id, app_info = NULL)

mod_title_server(id, ...)
}
\arguments{
\item{id}{The ID of the module.}

\item{app_info}{An optional list containing information about the application.
If left \code{NULL}, the function will use internally packaged app metadata.}
}
\value{
\itemize{
\item \code{mod_title_ui()}: A UI component for displaying the title.
\item \code{mod_title_server()}: Server logic for the title module.
}
}
\description{
A module to display the title of the application.

The module implements the following functions:
\itemize{
\item \code{mod_title_ui()}: User interface for displaying the title.
\item \code{mod_title_server()}: Server logic for the title module.
}
}
\details{
This module provides a simple user interface for displaying the title of the
application. The title is typically displayed in the top-left corner of the
application and may include a logo or other branding elements.
}
\section{App Info}{


The \code{app_info} parameter is a list containing the following elements:
\itemize{
\item \code{name}: The name of the application.
\item \code{version}: The version of the application.
\item \code{logo}: The path or URL to the application logo.
\item \code{symbol}: The path or URL to the application symbol or icon.
\item \code{repo_url}: The URL to the application's GitHub repository.
\item \code{docs_url}: The URL to the application's documentation.
}
}

\seealso{
\code{\link[=app_ui]{app_ui()}}
}
