% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_user_profile.R
\name{mod_user_profile}
\alias{mod_user_profile}
\alias{mod_user_profile_ui}
\alias{mod_user_profile_server}
\alias{mod_user_profile_demo}
\title{user_profile Shiny Module}
\usage{
mod_user_profile_ui(id)

mod_user_profile_server(id)

mod_user_profile_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_user_profile_ui()}: UI HTML Output.
\item \code{mod_user_profile_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for ...
\itemize{
\item \code{mod_user_profile_ui()}: User interface
\item \code{mod_user_profile_server()}: Server logic
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_user_profile_demo()
\dontshow{\}) # examplesIf}
}
