% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_list}
\alias{modify_list}
\title{Modify a List}
\usage{
modify_list(old, new)
}
\arguments{
\item{old, new}{lists}
}
\value{
updated list
}
\description{
This function contains the same functionality as \code{\link[utils:modifyList]{utils::modifyList()}} but
is more explicit in its purpose.
}
\seealso{
\code{\link[utils:modifyList]{utils::modifyList()}}
}
