% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{options}
\alias{options}
\title{gmhcommunities Options}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{gmhcommunities} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "gmhcommunities")

options::opt(x, default, env = "gmhcommunities")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{db.conn}{\describe{
Stores a database pooled connection object.\item{default: }{\preformatted{}}
\item{option: }{gmhcommunities.db.conn}
\item{envvar: }{R_GMHCOMMUNITIES_DB_CONN (evaluated if possible, raw string otherwise)}
}}

\item{db.user_id}{\describe{
Database user integer ID to use for query logging.\item{default: }{\preformatted{1}}
\item{option: }{gmhcommunities.db.user_id}
\item{envvar: }{R_GMHCOMMUNITIES_DB_USER_ID (evaluated if possible, raw string otherwise)}
}}

\item{entrata.base_url}{\describe{
Entrata API Base URL\item{default: }{\preformatted{get_entrata_config("base_url")}}
\item{option: }{gmhcommunities.entrata.base_url}
\item{envvar: }{R_GMHCOMMUNITIES_ENTRATA_BASE_URL (evaluated if possible, raw string otherwise)}
}}

\item{entrata.username}{\describe{
Entrata API Username\item{default: }{\preformatted{get_entrata_config("username")}}
\item{option: }{gmhcommunities.entrata.username}
\item{envvar: }{R_GMHCOMMUNITIES_ENTRATA_USERNAME (evaluated if possible, raw string otherwise)}
}}

\item{entrata.password}{\describe{
Entrata API Password\item{default: }{\preformatted{get_entrata_config("password")}}
\item{option: }{gmhcommunities.entrata.password}
\item{envvar: }{R_GMHCOMMUNITIES_ENTRATA_PASSWORD (evaluated if possible, raw string otherwise)}
}}

\item{entrata.user_agent}{\describe{
Entrata API User Agent\item{default: }{\preformatted{}}
\item{option: }{gmhcommunities.entrata.user_agent}
\item{envvar: }{R_GMHCOMMUNITIES_ENTRATA_USER_AGENT (evaluated if possible, raw string otherwise)}
}}

\item{entrata.timeout}{\describe{
Entrata API Timeout\item{default: }{\preformatted{30}}
\item{option: }{gmhcommunities.entrata.timeout}
\item{envvar: }{R_GMHCOMMUNITIES_ENTRATA_TIMEOUT (evaluated if possible, raw string otherwise)}
}}

\item{entrata.retry_on}{\describe{
Entrata API Retry On Status Codes\item{default: }{\preformatted{c(429, 500, 502, 503, 504)}}
\item{option: }{gmhcommunities.entrata.retry_on}
\item{envvar: }{R_GMHCOMMUNITIES_ENTRATA_RETRY_ON (evaluated if possible, raw string otherwise)}
}}

\item{entrata.max_retries}{\describe{
Entrata API Retry Max Attempts\item{default: }{\preformatted{3}}
\item{option: }{gmhcommunities.entrata.max_retries}
\item{envvar: }{R_GMHCOMMUNITIES_ENTRATA_MAX_RETRIES (evaluated if possible, raw string otherwise)}
}}

\item{entrata.request_id}{\describe{
Entrata Request ID\item{default: }{\preformatted{}}
\item{option: }{gmhcommunities.entrata.request_id}
\item{envvar: }{R_GMHCOMMUNITIES_ENTRATA_REQUEST_ID (evaluated if possible, raw string otherwise)}
}}

\item{entrata.verbosity}{\describe{
Entrata Verbosity\item{default: }{\preformatted{}}
\item{option: }{gmhcommunities.entrata.verbosity}
\item{envvar: }{R_GMHCOMMUNITIES_ENTRATA_VERBOSITY (evaluated if possible, raw string otherwise)}
}}

\item{entrata.default_headers}{\describe{
Entrata Default Headers\item{default: }{\preformatted{}}
\item{option: }{gmhcommunities.entrata.default_headers}
\item{envvar: }{R_GMHCOMMUNITIES_ENTRATA_DEFAULT_HEADERS (evaluated if possible, raw string otherwise)}
}}

\item{entrata.proxy}{\describe{
Entrata API Proxy\item{default: }{\preformatted{}}
\item{option: }{gmhcommunities.entrata.proxy}
\item{envvar: }{R_GMHCOMMUNITIES_ENTRATA_PROXY (evaluated if possible, raw string otherwise)}
}}

}
}

\seealso{
options getOption Sys.setenv Sys.getenv
}
