% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_parsers.R
\name{parse_floorplans_response}
\alias{parse_floorplans_response}
\title{Parse Entrata Floor Plans Response}
\usage{
parse_floorplans_response(resp)
}
\arguments{
\item{resp}{The Entrata API response object.}
}
\value{
A list containing the parsed floor plan data.
}
\description{
This function parses the response data from the Entrata API \verb{/properties} endpoint's
\code{getFloorPlans} method into various tibbles containing the response's floor plan-level
information and data.
}
\details{
The response is parsed into a merged tibble containing data across the
following sub-tables:
\itemize{
\item \code{floor_plans} - main floor plan information (e.g. name, unit counts)
\item \code{unit_types} - the unit types for the floor plans
\item \code{rooms} - number of bedrooms and bathrooms
\item \code{square_feet} - min, max, and average square feet
\item \code{market_rent} - min, max, and average market rent
\item \code{images} - image files associated with the floor plans
}
}
