% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_sys}
\alias{pkg_sys}
\title{Package System File}
\usage{
pkg_sys(...)
}
\arguments{
\item{...}{Path to file.}
}
\value{
Path to file.
}
\description{
This function is a wrapper for \code{\link[base:system.file]{base::system.file()}} that allows for
easier access to package files.
}
\seealso{
\code{\link[base:system.file]{base::system.file()}}
}
