% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{run_app}
\alias{run_app}
\title{Run Shiny App}
\usage{
run_app(ui = app_ui, server = app_server, cache = app_cache())
}
\arguments{
\item{ui}{Shiny App UI}

\item{server}{Shiny App Server}
}
\value{
\code{\link[shiny:shinyApp]{shiny::shinyApp()}}
}
\description{
Run Shiny App
}
\seealso{
Other App: 
\code{\link{app_assets}},
\code{\link{app_server}()},
\code{\link{app_ui}()}
}
\concept{App}
