% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{start_of_month}
\alias{start_of_month}
\title{Start of Month}
\usage{
start_of_month(as_of_date)
}
\arguments{
\item{as_of_date}{The date to use as the reference date. Default is today.}
}
\value{
The date of the first day of the month for the specified \code{as_of_date}.
}
\description{
Start of Month
}
