% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_leases.R
\name{summarize_leases}
\alias{summarize_leases}
\title{Summarize Entrata Lease Data}
\usage{
summarize_leases(lease_data)
}
\arguments{
\item{lease_data}{The lease data to summarize.}
}
\value{
A summarized tibble.
}
\description{
Summarizes the lease data by aggregating the scheduled charges.
}
