% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{utils_dates}
\alias{utils_dates}
\alias{get_weeks_of_year}
\alias{get_weekly_period}
\alias{get_weekly_period_start_date}
\alias{get_weekly_period_end_date}
\alias{get_leasing_period}
\alias{get_leasing_period_start_date}
\alias{get_leasing_period_end_date}
\alias{get_weeks_left_to_lease}
\title{Date Utilities}
\usage{
get_weeks_of_year(
  week_start = c("Monday", "Saturday", "Sunday"),
  year = lubridate::year(Sys.Date())
)

get_weekly_period(as_of_date = lubridate::today())

get_weekly_period_start_date(
  as_of_date = lubridate::today(),
  week_start = c("Monday", "Sunday")
)

get_weekly_period_end_date(as_of_date = lubridate::today())

get_leasing_period(as_of_date = lubridate::today())

get_leasing_period_start_date(
  as_of_date = lubridate::today(),
  format = "\%m/\%d/\%Y"
)

get_leasing_period_end_date(
  as_of_date = lubridate::today(),
  format = "\%m/\%d/\%Y"
)

get_weeks_left_to_lease(as_of_date = lubridate::today())
}
\arguments{
\item{week_start}{The day of the week to start the week on. Default is \code{"Monday"},
but can also be one of \code{"Saturday"} or \code{"Sunday"}.}

\item{year}{The year to get the weeks of. Default is the current year.}

\item{as_of_date}{The date to use as the reference date. Default is today.}

\item{format}{The format to return the date in. Default is \code{"\%m/\%d/\%Y"}
which is the format required for most dates when using the Entrata API.}
}
\value{
\itemize{
\item \code{get_weeks_of_year()}: A list of weeks in the year based off of the specified
week start and year.
\item \code{get_leasing_period_start_date()}: The start date of the leasing period.
\item \code{get_leasing_period_end_date()}: The end date of the leasing period.
\item \code{get_weeks_left_to_lease()}: The number of weeks left in the leasing period.
\item \code{get_weekly_period_start_date()}: The start date of the current week.
\item \code{get_weekly_period_end_date()}: The end date of the current week.
}
}
\description{
Date utility functions.
}
\details{
The following functions are included:
\itemize{
\item \code{get_weeks_of_year()}: Get the (leasing) weeks of the year based off of
the specified reporting cycle (i.e. what is the start of the week based from)
and the current year.
\item \code{get_leasing_period_start_date()}: Get the start date of the leasing period
based off of the specified date.
\item \code{get_leasing_period_end_date()}: Get the end date of the leasing period
based off of the specified date.
\item \code{get_weeks_left_to_lease()}: Get the number of weeks left in the leasing period
based off of the specified date.
\item \code{get_weekly_period_start_date()}: Get the start date of the current week.
\item \code{get_weekly_period_end_date()}: Get the end date of the current week.
}
}
\seealso{
\code{\link[=entrata_properties]{entrata_properties()}}, \code{\link[=entrata_reports]{entrata_reports()}}, \code{\link[=entrata_config]{entrata_config()}}
}
\concept{Dates}
\concept{Utilities}
