% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{with_loader}
\alias{with_loader}
\title{Add Loading Spinners to Outputs}
\usage{
with_loader(output, ...)
}
\arguments{
\item{output}{The output element to add the loading spinner to.}

\item{...}{
  Arguments passed on to \code{\link[shinycustomloader:withLoader]{shinycustomloader::withLoader}}
  \describe{
    \item{\code{ui_element}}{A UI element that should be wrapped with a loader when the corresponding output is being calculated.}
    \item{\code{type}}{The type of loader to use. Possible values are image, html or text.}
    \item{\code{loader}}{The name of the loader. The built in options are dnaspin, dancingtree, pacman and walkingcow.
For custom html option, the name of the css and html file have to be the same and input must be without the extension.
For custom image option, the name must include the extension of the file.
For custom text option, loader must be a list of marquee objects.}
    \item{\code{proxy.height}}{If the output doesn't specify the output height, you can set a proxy height. It defaults to 400px for outputs with undefined height.}
  }}
}
\value{
The output element with the loading spinner added.
}
\description{
This function adds a loading spinner to a Shiny output element via
\code{\link[shinycustomloader:withLoader]{shinycustomloader::withLoader()}}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(shiny)

ui <- fluidPage(
  with_loader(
    plotOutput("plot")
  ),
  plotOutput("plot") |>
    with_loader()
  )
)

server <- function(input, output, session) {

  output$plot <- renderPlot({
    plot(1:10)
  })

  outputOptions(output, "plot", suspendWhenHidden = FALSE)

}

shinyApp(ui = ui, server = server)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[shinycustomloader:withLoader]{shinycustomloader::withLoader()}}
}
