% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{with_tooltip}
\alias{with_tooltip}
\title{Add Tooltips to Inputs}
\usage{
with_tooltip(input, tooltip_text, placement = "right")
}
\arguments{
\item{input}{The input element to add the tooltip to.}

\item{tooltip_text}{The text to display in the tooltip.}

\item{placement}{The placement of the tooltip. Default is "right".}
}
\value{
The input element with the tooltip added.
}
\description{
This function adds a tooltip to a Shiny input element via \code{\link[bslib:tooltip]{bslib::tooltip()}}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(shiny)
library(bslib)
library(bsicons)

ui <- fluidPage(
  with_tooltip(
    textInput("text", "Text Input"),
    "This is a tooltip."
    ),
    # or pipe the input into the function
    textInput("text", "Text Input") |>
      with_tooltip("This is a tooltip.")
)

shinyApp(ui = ui, server = function(input, output, session) {})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[bslib:tooltip]{bslib::tooltip()}}
}
